replace private details/metadata in email content
when replying to an email many clients copy the header (from/to/subject) into the body.
because the relay system needs to send the email in the name of the domain
this results in private information being added to the email.
Additionally it will confuse the recipient because his email will appear
to be sent from the domain to some private email (as opposed to "from him to the domain")

This now simply string replaces the content if the content matches a predefined regex pattern